Split the capability manager into multiple files, one per shim.
