+++ capability_manager_shim_da_device_compose.go
@@ -1,0 +1,15 @@
+package zda
+
+import "github.com/shimmeringbee/da"
+
+type composeDADeviceShim struct {
+	gateway da.Gateway
+}
+
+func (s *composeDADeviceShim) Compose(zdaDevice Device) da.Device {
+	return da.BaseDevice{
+		DeviceGateway:      s.gateway,
+		DeviceIdentifier:   zdaDevice.Identifier,
+		DeviceCapabilities: zdaDevice.Capabilities,
+	}
+}
