+++ capability_manager_shim_event_sender.go
@@ -1,0 +1,9 @@
+package zda
+
+type daEventSenderShim struct {
+	eventSender eventSender
+}
+
+func (s *daEventSenderShim) Send(e interface{}) {
+	s.eventSender.sendEvent(e)
+}
