+++ capability_manager_test.go
@@ -3,12 +3,9 @@ package zda
 import (
 	"context"
 	"github.com/shimmeringbee/da"
-	"github.com/shimmeringbee/zigbee"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
-	"sync"
 	"testing"
-	"time"
 )
 
 func TestCapabilityManager_Add(t *testing.T) {
@@ -225,351 +222,5 @@ func TestCapabilityManager_initSupervisor_FetchCapability(t *testing.T) {
 		s := m.initSupervisor()
 
 		assert.Equal(t, &m, s.FetchCapability())
-	})
-}
-
-func TestCapabilityManager_initSupervisor_ManageDeviceCapabilities(t *testing.T) {
-	t.Run("provides an implementation that calls the gateway capabilities for add capability", func(t *testing.T) {
-		mdcm := &mockDeviceCapabilityManager{}
-		defer mdcm.AssertExpectations(t)
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		zdaDevice := Device{
-			Identifier: addr,
-		}
-		c := da.Capability(0x0001)
-
-		mdcm.On("AddCapability", addr, c)
-
-		m := CapabilityManager{deviceCapabilityManager: mdcm}
-		s := m.initSupervisor()
-
-		s.ManageDeviceCapabilities().Add(zdaDevice, c)
-	})
-
-	t.Run("provides an implementation that calls the gateway capabilities for add capability", func(t *testing.T) {
-		mdcm := &mockDeviceCapabilityManager{}
-		defer mdcm.AssertExpectations(t)
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		zdaDevice := Device{
-			Identifier: addr,
-		}
-		c := da.Capability(0x0001)
-
-		mdcm.On("RemoveCapability", addr, c)
-
-		m := CapabilityManager{deviceCapabilityManager: mdcm}
-		s := m.initSupervisor()
-
-		s.ManageDeviceCapabilities().Remove(zdaDevice, c)
-	})
-}
-
-func TestCapabilityManager_initSupervisor_DAEventSender(t *testing.T) {
-	t.Run("provides an implementation that calls the gateway capabilities for add capability", func(t *testing.T) {
-		mes := &mockEventSender{}
-		defer mes.AssertExpectations(t)
-
-		mes.On("sendEvent", mock.Anything)
-
-		m := CapabilityManager{eventSender: mes}
-		s := m.initSupervisor()
-
-		s.DAEventSender().Send(nil)
-	})
-}
-
-func TestCapabilityManager_initSupervisor_CDADevice(t *testing.T) {
-	t.Run("provides an implementation that creates a da device from a zda device", func(t *testing.T) {
-		zgw := &ZigbeeGateway{}
-
-		m := CapabilityManager{gateway: zgw}
-		s := m.initSupervisor()
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		capabilities := []da.Capability{0x0001}
-
-		zdaDevice := Device{
-			Identifier:   addr,
-			Capabilities: capabilities,
-			Endpoints:    nil,
-		}
-
-		d := s.ComposeDADevice().Compose(zdaDevice)
-
-		assert.Equal(t, addr, d.Identifier())
-		assert.Equal(t, capabilities, d.Capabilities())
-		assert.Equal(t, zgw, d.Gateway())
-	})
-}
-
-func TestCapabilityManager_initSupervisor_DeviceLookup(t *testing.T) {
-	t.Run("returns false if gateway doesn't match", func(t *testing.T) {
-		zgw := &ZigbeeGateway{}
-
-		m := CapabilityManager{gateway: zgw, nodeTable: newNodeTable()}
-		s := m.initSupervisor()
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		daDevice := da.BaseDevice{DeviceIdentifier: addr}
-
-		_, ok := s.DeviceLookup().ByDA(daDevice)
-		assert.False(t, ok)
-	})
-
-	t.Run("returns false if gateway does match, but isn't found in the node table", func(t *testing.T) {
-		zgw := &ZigbeeGateway{}
-
-		m := CapabilityManager{gateway: zgw, nodeTable: newNodeTable()}
-		s := m.initSupervisor()
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		daDevice := da.BaseDevice{DeviceIdentifier: addr, DeviceGateway: zgw}
-
-		_, ok := s.DeviceLookup().ByDA(daDevice)
-		assert.False(t, ok)
-	})
-
-	t.Run("returns true if gateway does match and is found, device details match", func(t *testing.T) {
-		zgw := &ZigbeeGateway{}
-
-		m := CapabilityManager{gateway: zgw, nodeTable: newNodeTable()}
-		s := m.initSupervisor()
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		daDevice := da.BaseDevice{DeviceIdentifier: addr, DeviceGateway: zgw}
-
-		iN, _ := m.nodeTable.createNode(addr.IEEEAddress)
-		iD, _ := m.nodeTable.createDevice(addr)
-		iD.capabilities = []da.Capability{0x0001}
-
-		iN.endpoints = []zigbee.Endpoint{0x01, 0x02}
-		iN.endpointDescriptions = map[zigbee.Endpoint]zigbee.EndpointDescription{
-			0x01: {
-				Endpoint: 0x01,
-			},
-			0x02: {
-				Endpoint: 0x02,
-			},
-		}
-
-		iD.endpoints = []zigbee.Endpoint{0x02}
-
-		expectedEndpoints := map[zigbee.Endpoint]zigbee.EndpointDescription{
-			0x02: {
-				Endpoint: 0x02,
-			},
-		}
-
-		d, ok := s.DeviceLookup().ByDA(daDevice)
-		assert.True(t, ok)
-
-		assert.Equal(t, iD.capabilities, d.Capabilities)
-		assert.Equal(t, addr, d.Identifier)
-		assert.Equal(t, expectedEndpoints, d.Endpoints)
-	})
-}
-
-func TestCapabilityManager_initSupervisor_Poller(t *testing.T) {
-	t.Run("calls add on the parent poller with identifier", func(t *testing.T) {
-		mockPoller := &mockPoller{}
-		defer mockPoller.AssertExpectations(t)
-
-		m := CapabilityManager{poller: mockPoller}
-		s := m.initSupervisor()
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		zdaDevice := Device{Identifier: addr}
-
-		interval := 10 * time.Millisecond
-
-		mockPoller.On("Add", addr, interval, mock.Anything)
-
-		cFn := s.Poller().Add(zdaDevice, interval, func(ctx context.Context, device Device) bool {
-			return true
-		})
-
-		assert.NotNil(t, cFn)
-	})
-
-	t.Run("when capability provided function is called a populated zda is provided", func(t *testing.T) {
-		mockPoller := &mockPoller{}
-		defer mockPoller.AssertExpectations(t)
-
-		m := CapabilityManager{poller: mockPoller}
-		s := m.initSupervisor()
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		zdaDevice := Device{Identifier: addr}
-
-		interval := 10 * time.Millisecond
-
-		mockPoller.On("Add", addr, interval, mock.Anything)
-
-		called := false
-
-		_, _, iDev := generateNodeTableWithData(1)
-
-		s.Poller().Add(zdaDevice, interval, func(ctx context.Context, device Device) bool {
-			assert.Equal(t, iDev[0].generateIdentifier(), device.Identifier)
-			called = true
-			return true
-		})
-
-		wrappedFn, ok := mockPoller.Calls[0].Arguments[2].(func(context.Context, *internalDevice) bool)
-		assert.True(t, ok)
-
-		ret := wrappedFn(context.TODO(), iDev[0])
-
-		assert.True(t, called)
-		assert.True(t, ret)
-	})
-
-	t.Run("when poller is cancelled the wrapper returns false without calling the wrapped function", func(t *testing.T) {
-		mockPoller := &mockPoller{}
-		defer mockPoller.AssertExpectations(t)
-
-		m := CapabilityManager{poller: mockPoller}
-		s := m.initSupervisor()
-
-		addr := IEEEAddressWithSubIdentifier{IEEEAddress: zigbee.GenerateLocalAdministeredIEEEAddress(), SubIdentifier: 0x11}
-		zdaDevice := Device{Identifier: addr}
-
-		interval := 10 * time.Millisecond
-
-		mockPoller.On("Add", addr, interval, mock.Anything)
-
-		_, _, iDev := generateNodeTableWithData(1)
-
-		cFn := s.Poller().Add(zdaDevice, interval, func(ctx context.Context, device Device) bool {
-			t.Fatalf("should not have run wrapper")
-			return true
-		})
-
-		wrappedFn, ok := mockPoller.Calls[0].Arguments[2].(func(context.Context, *internalDevice) bool)
-		assert.True(t, ok)
-
-		cFn()
-		ret := wrappedFn(context.TODO(), iDev[0])
-
-		assert.False(t, ret)
-	})
-}
-
-func TestCapabilityManager_deviceAddedCallback(t *testing.T) {
-	t.Run("calls any capabilities that have the device management interface", func(t *testing.T) {
-		m := CapabilityManager{
-			capabilityByFlag:    map[da.Capability]interface{}{},
-			capabilityByKeyName: map[string]PersistableCapability{},
-		}
-
-		f := da.Capability(0x0000)
-		kN := "KEY"
-
-		mC := &mockCapability{}
-		mC.On("Capability").Return(f)
-		mC.On("KeyName").Return(kN)
-
-		m.Add(mC)
-
-		node := &internalNode{
-			mutex:                &sync.RWMutex{},
-			endpointDescriptions: map[zigbee.Endpoint]zigbee.EndpointDescription{},
-		}
-		device := &internalDevice{
-			node:  node,
-			mutex: &sync.RWMutex{},
-		}
-		zdaDevice := Device{
-			Endpoints: map[zigbee.Endpoint]zigbee.EndpointDescription{},
-		}
-
-		ctx := context.TODO()
-
-		mC.On("AddedDevice", ctx, zdaDevice).Return(nil)
-		defer mC.AssertExpectations(t)
-
-		err := m.deviceAddedCallback(ctx, internalDeviceAdded{device: device})
-		assert.NoError(t, err)
-
-	})
-}
-
-func TestCapabilityManager_deviceRemoveCallback(t *testing.T) {
-	t.Run("calls any capabilities that have the device management interface", func(t *testing.T) {
-		m := CapabilityManager{
-			capabilityByFlag:    map[da.Capability]interface{}{},
-			capabilityByKeyName: map[string]PersistableCapability{},
-		}
-
-		f := da.Capability(0x0000)
-		kN := "KEY"
-
-		mC := &mockCapability{}
-		mC.On("Capability").Return(f)
-		mC.On("KeyName").Return(kN)
-
-		m.Add(mC)
-
-		node := &internalNode{
-			mutex:                &sync.RWMutex{},
-			endpointDescriptions: map[zigbee.Endpoint]zigbee.EndpointDescription{},
-		}
-		device := &internalDevice{
-			node:  node,
-			mutex: &sync.RWMutex{},
-		}
-		zdaDevice := Device{
-			Endpoints: map[zigbee.Endpoint]zigbee.EndpointDescription{},
-		}
-
-		ctx := context.TODO()
-
-		mC.On("RemovedDevice", ctx, zdaDevice).Return(nil)
-		defer mC.AssertExpectations(t)
-
-		err := m.deviceRemovedCallback(ctx, internalDeviceRemoved{device: device})
-		assert.NoError(t, err)
-
-	})
-}
-
-func TestCapabilityManager_deviceEnumeratedCallback(t *testing.T) {
-	t.Run("calls any capabilities that have the device management interface", func(t *testing.T) {
-		m := CapabilityManager{
-			capabilityByFlag:    map[da.Capability]interface{}{},
-			capabilityByKeyName: map[string]PersistableCapability{},
-		}
-
-		f := da.Capability(0x0000)
-		kN := "KEY"
-
-		mC := &mockCapability{}
-		mC.On("Capability").Return(f)
-		mC.On("KeyName").Return(kN)
-
-		m.Add(mC)
-
-		node := &internalNode{
-			mutex:                &sync.RWMutex{},
-			endpointDescriptions: map[zigbee.Endpoint]zigbee.EndpointDescription{},
-		}
-		device := &internalDevice{
-			node:  node,
-			mutex: &sync.RWMutex{},
-		}
-		zdaDevice := Device{
-			Endpoints: map[zigbee.Endpoint]zigbee.EndpointDescription{},
-		}
-
-		ctx := context.TODO()
-
-		mC.On("EnumerateDevice", ctx, zdaDevice).Return(nil)
-		defer mC.AssertExpectations(t)
-
-		err := m.deviceEnumeratedCallback(ctx, internalDeviceEnumeration{device: device})
-		assert.NoError(t, err)
-
 	})
 }
